+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new PeriodSetupNode primitive
 - new FactorAnalysisNode primitive
 - new RiskRAppetiteNode primitive
+- default label for nodes
+- AbstractNode class
+- NodeWithDefaultLabel class
 
 ## [0.2.0]